requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
DateTime==5.5
DAWG2-Python==0.9.0
et_xmlfile==2.0.0
fastapi==0.115.14
fasttext-wheel==0.9.2
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
portalocker==2.10.1
protobuf==6.31.1
pure_eval==0.2.3
pybind11==2.13.6
pydantic==2.11.7
pydantic_core==2.33.2
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
qdrant-client==1.14.3
regex==2024.11.6
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
zope.interface==7.2
